+++ src/types/PaymentSheet.ts
@@ -1,5 +1,11 @@
 import type { BillingDetails, AddressDetails } from './Common';
 import type { CartSummaryItem } from './ApplePay';
+import type {
+  ButtonType,
+  RecurringPaymentRequest,
+  AutomaticReloadPaymentRequest,
+  MultiMerchantRequest,
+} from './PlatformPay';
 
 export type SetupParams = ClientSecretParams & {
   /** Your customer-facing business name. On Android, this is required and cannot be an empty string. */
@@ -57,9 +63,26 @@ export type ApplePayParams = {
   /**
    * An array of CartSummaryItem item objects that summarize the amount of the payment. If you're using a SetupIntent
    * for a recurring payment, you should set this to display the amount you intend to charge. */
-  paymentSummaryItems?: CartSummaryItem[];
-  // TODO: Uncomment when https://github.com/stripe/stripe-react-native/pull/1164 lands
-  // buttonType: ButtonType
+  cartItems?: CartSummaryItem[];
+  /** Sets the text displayed by the call to action button in the Apple Pay sheet. */
+  buttonType?: ButtonType;
+  /** A typical request is for a one-time payment. To support different types of payment requests, include a PaymentRequestType. Only supported on iOS 16 and up. */
+  request?:
+    | RecurringPaymentRequest
+    | AutomaticReloadPaymentRequest
+    | MultiMerchantRequest;
+  /** Callback function for setting the order details (retrieved from your server) to give users the
+   * ability to track and manage their purchases in Wallet. Stripe calls your implementation after the
+   * payment is complete, but before iOS dismisses the Apple Pay sheet. You must call the `completion`
+   * function, or else the Apple Pay sheet will hang. */
+  setOrderTracking?: (
+    completion: (
+      orderIdentifier: string,
+      orderTypeIdentifier: string,
+      authenticationToken: string,
+      webServiceUrl: string
+    ) => void
+  ) => void;
 };
 
 export type GooglePayParams = {